+++ src/Hyperion.Tests/Generator/CrossFrameworkInitializer.cs
@@ -1,10 +1,30 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
 
 namespace Hyperion.Tests.Generator
 {
     public static class CrossFrameworkInitializer
     {
         public const string DefaultOutputPath = "../../../testfiles";
+
+        public static CrossFrameworkMixedClass InitMixed()
+        {
+            return new CrossFrameworkMixedClass
+            {
+                Name = "Cookie",
+                Sound = "Bark",
+                FriendType = typeof(CrossFrameworkClass),
+                Color = Color.Blue,
+                Point = new Point(10, 10),
+                PointF = new PointF(10, 10),
+                Rectangle = new Rectangle(10, 10, 10, 10),
+                RectangleF = new RectangleF(10, 10, 10, 10),
+                Size = new Size(10, 10),
+                SizeF = new SizeF(10, 10),
+                Data = Init()
+            };
+        }
 
         public static CrossFrameworkClass Init()
         {