+++ src/Hyperion/SerializerFactories/DictionarySerializerFactory.cs
@@ -11,6 +11,7 @@ using System;
 using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reflection;
 using Hyperion.Extensions;
@@ -39,26 +40,63 @@ namespace Hyperion.SerializerFactories
             var preserveObjectReferences = serializer.Options.PreserveObjectReferences;
             var ser = new ObjectSerializer(type);
             typeMapping.TryAdd(type, ser);
-            var elementSerializer = serializer.GetSerializerByType(typeof (DictionaryEntry));
+            var dictionaryTypes = GetKeyValuePairType(type);
+            var elementSerializer = serializer.GetSerializerByType(dictionaryTypes.KeyValuePairType);
 
             ObjectReader reader = (stream, session) =>
             {
-                throw new NotSupportedException("Generic IDictionary<TKey,TValue> are not yet supported");
-#pragma warning disable CS0162 // Unreachable code detected
-                var instance = Activator.CreateInstance(type);
-#pragma warning restore CS0162 // Unreachable code detected
+                object instance;
+
+                void ReadDictionaryKeyValuePairs(object dictionaryInstance)
+                {
+                    var count = stream.ReadInt32(session);
+                    for (var i = 0; i < count; i++)
+                    {
+                        var entry = stream.ReadObject(session); // KeyValuePair<TKey, TValue>
+
+                        // Get entry.Key and entry.Value
+                        var key = dictionaryTypes.KeyValuePairType.GetProperty(nameof(KeyValuePair<object, object>.Key)).GetValue(entry, null);
+                        var value = dictionaryTypes.KeyValuePairType.GetProperty(nameof(KeyValuePair<object, object>.Value)).GetValue(entry, null);
+
+                        // Same as: instance.Add(key, value)
+                        dictionaryTypes.DictionaryInterfaceType
+                            .GetMethod(nameof(IDictionary<object, object>.Add), new[] { dictionaryTypes.KeyType, dictionaryTypes.ValueType })
+                            .Invoke(dictionaryInstance, new[] { key, value });
+                    }
+                }
+
+                #region Special case for ReadOnlyDictionary
+                // Special case for ReadOnlyDictionary since ReadOnlyDictionary
+                // does not have a parameterless constructor
+                var genericReadOnlyDictionary = typeof(ReadOnlyDictionary<,>);
+                var readOnlyDictionaryType = 
+                    genericReadOnlyDictionary.MakeGenericType(dictionaryTypes.KeyType, dictionaryTypes.ValueType);
+                if (type.Equals(readOnlyDictionaryType))
+                {
+                    var genericDictionary = typeof(Dictionary<,>);
+                    var genericDictionaryType = genericDictionary.MakeGenericType(dictionaryTypes.KeyType, dictionaryTypes.ValueType);
+                    var dictionary = Activator.CreateInstance(genericDictionaryType); // normal dictionary
+
+                    ReadDictionaryKeyValuePairs(dictionary);
+                    instance = Activator.CreateInstance(type, dictionary); // IDictionary<TKey, TValue>
+                    if (preserveObjectReferences) session.TrackDeserializedObject(instance);
+                    return instance;
+                }
+                #endregion
+
+                try
+                {
+                    instance = Activator.CreateInstance(type, true); // IDictionary<TKey, TValue>
+                } catch(Exception) {
+                    instance = Activator.CreateInstance(type); // IDictionary<TKey, TValue>
+                }
+
                 if (preserveObjectReferences)
                 {
                     session.TrackDeserializedObject(instance);
                 }
-                var count = stream.ReadInt32(session);
-                var entries = new DictionaryEntry[count];
-                for (var i = 0; i < count; i++)
-                {
-                    var entry = (DictionaryEntry) stream.ReadObject(session);
-                    entries[i] = entry;
-                }
-                //TODO: populate dictionary
+                ReadDictionaryKeyValuePairs(instance);
+                
                 return instance;
             };
 
@@ -68,19 +106,44 @@ namespace Hyperion.SerializerFactories
                 {
                     session.TrackSerializedObject(obj);
                 }
-                var dict = obj as IDictionary;
+                
+                var dict = obj as IEnumerable; // IDictionary<T, V> is IEnumerable<KeyValuePair<T, V>>
+                var count = dict.Cast<object>().Count();
                 // ReSharper disable once PossibleNullReferenceException
-                Int32Serializer.WriteValueImpl(stream,dict.Count,session);
+                Int32Serializer.WriteValueImpl(stream, count, session);
                 foreach (var item in dict)
                 {
-                    stream.WriteObject(item, typeof (DictionaryEntry), elementSerializer,
-                        serializer.Options.PreserveObjectReferences, session);
-                    // elementSerializer.WriteValue(stream,item,session);
+                    stream.WriteObject(item, dictionaryTypes.KeyValuePairType, elementSerializer, serializer.Options.PreserveObjectReferences, session);
                 }
             };
             ser.Initialize(reader, writer);
             
             return ser;
+        }
+
+        private GenericDictionaryTypes GetKeyValuePairType(Type dictImplementationType)
+        {
+            var dictInterface = dictImplementationType
+                .GetInterfaces()
+                .Where(i => i.GetTypeInfo().IsGenericType)
+                .First(i => i.GetGenericTypeDefinition() == typeof(IDictionary<,>));
+            var keyType = dictInterface.GetGenericArguments()[0];
+            var valueType = dictInterface.GetGenericArguments()[1];
+            return new GenericDictionaryTypes()
+            {
+                KeyType = keyType,
+                ValueType = valueType,
+                KeyValuePairType = typeof(KeyValuePair<,>).MakeGenericType(keyType, valueType),
+                DictionaryInterfaceType = typeof(IDictionary<,>).MakeGenericType(keyType, valueType)
+            };
+        }
+
+        class GenericDictionaryTypes
+        {
+            public Type KeyType { get; set; }
+            public Type ValueType { get; set; }
+            public Type KeyValuePairType { get; set; }
+            public Type DictionaryInterfaceType { get; set; }
         }
     }
 }