+++ src/Hyperion/ValueSerializers/DoubleSerializer.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.IO;
+using Hyperion.Extensions;
 
 namespace Hyperion.ValueSerializers
 {
@@ -30,7 +31,7 @@ namespace Hyperion.ValueSerializers
 
         public static double ReadValueImpl(Stream stream, byte[] bytes)
         {
-            stream.Read(bytes, 0, Size);
+            stream.ReadFull(bytes, 0, Size);
             return BitConverter.ToDouble(bytes, 0);
         }