+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
-### 0.9.11 November 13 2019 ####
+### 0.10.2 June 30 2021 ####
+* [Update Akka version to 1.4.21](https://github.com/akkadotnet/akka.net/releases/tag/1.4.21)
+* [Add exception rethrow to help with debugging](https://github.com/akkadotnet/Hyperion/pull/229)
 
-[Hyperion now targets .NET 4.5 again so it can be included inside Akka.NET v1.3.* releases](https://github.com/akkadotnet/Hyperion/pull/141).
+### 0.10.1 April 20 2021 ####
+* [Fix SerializerOptions constructor backward compatibility issue with Akka.NET](https://github.com/akkadotnet/Hyperion/pull/214)